calibration/requirements.txt
matplotlib
pandas
pyserial
tqdm
scikit-learn
rocketlogger
pyyaml